Allow https OR http download types
The topcat Vagrant build only configures http downloads for LILS, not
https. The script assumes https, tries to select it, and dies if it is
not an option. This change tests for both http and https, and will use
http for the download if https is not available. If neither is
available, it reports and skips the download cart tests.
I'm not convinced that absence of https should be considered a failure,
though perhaps absence of both http and https should!