Get the downloaded datafile name from the page content
Replaced a hard-wired datafile name by a selector that gets (or attempts
to get) the datafile name from the content. This uses an XPath
expression that selects the first span under the first div with a
particular style class; not sure how robust that will be.
This removes the requirement that the ICAT test data includes datasets
that contain a datafile (indeed, whose "first" datafile is) called
"Datafile 1".